+++ api/authorization/v1alpha1/applyconfiguration/ssa/patchhelper.go
@@ -6,6 +6,7 @@ package ssa
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"slices"